// src/api/panelsdesu/models/getSearchSearchMedia200PanelsItem.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * PanelsDesu API
 * The official API for PanelsDesu
 * OpenAPI spec version: 1.0.0
 */

export type GetSearchSearchMedia200PanelsItem = {
  created_at?: string;
  description?: string;
  height?: number;
  id?: string;
  image_url?: string;
  manga_id?: string;
  ocr_text?: string;
  page_id?: string;
  phash?: string;
  width?: number;
  x?: number;
  y?: number;
};